refactor: l2 banners code cleanup
