--- shared/banners/index.ts
@@ -1,4 +1,0 @@
-export * from './curve';
-export * from './modal-pool-banners';
-export * from './l2-low-fee';
-export * from './l2-swap';
