--- shared/banners/l2-oneinch/index.ts
@@ -1,1 +1,0 @@
-export * from './l2-oneinch';
