--- shared/banners/l2-swap/index.ts
@@ -1,1 +1,0 @@
-export * from './l2-swap';
