+++ shared/banners/l2-wsteth/l2-wsteth.tsx
@@ -1,13 +1,5 @@
 import { useCallback } from 'react';
-import {
-  Banner,
-  L2Icons,
-  LinkButton,
-  TextContent,
-  OverlayLink,
-  FooterWrapper,
-} from './styles';
-import { L2_DISCOVERY_LINK } from '../l2-banner';
+import { L2Banner } from '../l2-banner';
 import { MATOMO_CLICK_EVENTS } from 'config';
 import { trackEvent } from '@lidofinance/analytics-matomo';
 
@@ -23,22 +15,17 @@ export const L2Wsteth = ({ matomoEventLink }: L2WstethProps) => {
     [matomoEventLink],
   );
   return (
-    <Banner data-testid="L2wstETHbanner">
-      <TextContent>
-        Enjoy <b>lower gas</b> fees and <b>DeFi opportunities</b> using wstETH
-        across Arbitrum, Optimism, Base, zkSync, Mantle and Linea.
-      </TextContent>
-      <FooterWrapper>
-        <L2Icons />
-        <OverlayLink
-          href={L2_DISCOVERY_LINK}
-          onClick={linkClickHandler}
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <LinkButton data-testid="l2WSstethlearnMore">Learn More</LinkButton>
-        </OverlayLink>
-      </FooterWrapper>
-    </Banner>
+    <L2Banner
+      testidWrap="L2wstETHbanner"
+      testidButton="l2WSstethlearnMore"
+      buttonText="Learn More"
+      onClickButton={linkClickHandler}
+      text={
+        <>
+          Enjoy <b>lower gas</b> fees and <b>DeFi opportunities</b> using wstETH
+          across Arbitrum, Optimism, Base, zkSync, Mantle and Linea.
+        </>
+      }
+    />
   );
 };
